# Byte Align bstream writes