--- a/mdata/chunk/chunk.go
+++ b/mdata/chunk/chunk.go
@@ -10,12 +10,8 @@ import (
 )
 
 // Chunk is a chunk of data. not concurrency safe.
-// last check that the methods are being called safely by Dieter on 20/11/2018
-// checked: String, Push, Finish, Encode and properties Series, NumPoints, First
+// last check that all properties and methods are being accessed safely by Dieter on 2021-10-06
 // for the most part, confirming serialized access is easy by tracking all callers/references.
-// The main exception is the ChunkWriteRequest mechanism. any CWR created is processed
-// asynchronously, and chunk properties will be read (the series.T0, chunk.Encode(), etc)
-// But it can be proven that we only call NewChunkWriteRequest() on chunks that are no longer being modified.
 type Chunk struct {
 	Series    tsz.SeriesLong
 	NumPoints uint32