--- a/mdata/chunk/tsz/fuzz.go
+++ b/mdata/chunk/tsz/fuzz.go
@@ -1,3 +1,4 @@
+//go:build gofuzz
 // +build gofuzz
 
 package tsz